safari-534.53-branch/LayoutTests/fast/loader/file-protocol-fragment.html

<html>
<head>
<script>
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.dumpResourceLoadCallbacks();
        if (layoutTestController.setSerializeHTTPLoads)
            layoutTestController.setSerializeHTTPLoads();
    } else {
        document.write ('This test must be run through DumpRenderTree.')
    }
</script>
<link rel=stylesheet href="resources/stylesheet.css#a">
<link rel=stylesheet href="resources/stylesheet.css#b">
<body>
This tests that file: URL are loaded separately even if they differ by the fragment only.
</div>
</html>
